Add more support to GT for subdivision creases.
When refining to subdivision meshes, build the crease/corner tags if
there are crease weights.

When buidling geometry, if there are crease/corner weights, build the
appropriate vertex attribute and set the values accordingly.

In the Alembic ROP, don't output creaseweight attributes when saving
subdivision surfaces.  Instead, generate the creases on the
subdivision primitive in refinement.

In the GT procedural, convert the creases to creaseweight attributes
on the geometry when rendering subdivision surfaces.