Properly compute velocity bounds for various packed primitive types.
In VRAY, use the packed primitive velocity range method to determine
this rather than just looking for the attribute.  This picks up
velocity in nested GT primitives and is more accurate.

It's not 100% accurate.  For PackedDisk, there's a case where if there
are nested packed primitives in the disk file, we don't properly
interrogate the velocity bounds of the disk file.