Fix a number of issues with subd rendering, particularly with uniform subdivision. Uniform subd now has the same optimization to prevent rebuilding the OSD mesh when the topology and vertex attributes haven't changed. Fix bug with uniform subd where the subd would seem to turn off when re-rendering geometry with unchanged topology (which would happen when diving from objects into SOPs or changing the selection).
Changed the code for alembic topology comparisons. If the geometry is animating the topology attribute is set to -1 always, which is considered as unmatchable by the GR code (much like negaitve Data Array ids). For non-animated geometry, the topology value is a hash of the object path and file path. I dug into this because I had two alembic files with the same object paths but different topology which was causing a crash because the file path wasn't being taken into account. This way the GR code doesn't need to explicitly check the object path or file path to detect topology changes. This let me also get rid of the __object_path attribute created on alembic primitives as a way to get the intrinsic path information from GEO into GT.

Attempt to fix AMD issue with drawing shadows of subdivided surfaces.