Add new option to flatten visibility in the Alembic archive (enabled by default).  This is required to properly reproduce Alembic visibility (which is handled by the SOP properly).
In a case where you have:

    parent:  Animated visibility
	child1:  Deferred visibility
	child2:  Visible

child1 will inherit visibility from the parent, but child2 will be
visible regardless of the parent's visibility.

To get correct visibility, we need to do evaluation of the visibility
on the leaf geometry nodes.  However, to properly export the Alembic,
we need to keep the information on the subnet nodes, which is why the
parameter exists.