Provide IPR updating of style sheets, including triggering of an IPR refresh whenever an internally defined style sheet is modified. There were a lot of hole in VRAY in terms of being able to properly propagate style sheet changes from the top level down into all affected instances. Supporting this required tracking not just the styler for each VRAY_Instance, but the parent style and the STY_Subject used to create the styler. Updating a top level style sheet can then be passed down to each top level instance, which can pass it to it child instances without caring how the intervening STY_Subject was created (it may be an agent shape subject, primitive subject, object subject, or whatever).
The tracking of changes to internal stylesheets in SOHO is quite non-standard
because I couldn't figure out SOHO_Objects well enough to shoehorn the global
style sheets into that structure. So internal style sheets are tracked with
a DEP_MicroNode, and a map of style names to version ids so we know when a
style changes. We then use a custom property to pass the list of dirty style
sheets to the SOHO update code.