Added an Object Exclude parameter to the Alembic SOP. This parm is like the Object Path parm in terms of how the value is formatted, but indicates Alembic object which should not be loaded (weeds out parts of the Object Path).
The Object Path and Exclude Path parms both have a tree browser (which is
now a button with an icon instead of the word "Pick"). Both parms also have
a python action script button which starts an Alembic viewport selection.
In the case of Object Path, both Object Path and Exclude Path are cleared
before picking begins. For Exclude Path, the Object Path parm is left so
you can pick only Alembic prims included in the Object Path to exclude
(which is the whole point).