Changed VGEO_Detail to contain a GU_ConstDetailHandle to hold the geometry.  This allows people to pass in an existing detail to be shared by VGEO.
This required several significant changes to the way VRAY_Procedural
handled geometry.  The new interfaces allow geometry to be shared
properly.  Several deprecated methods have been removed.

This means that packed disk primitives no longer have to duplicate the
geometry, but instead can share it.  This results in a savings of 1.3
GB when rendering grunt as a packed diks (for example).

Also performed some additional cleanup:
    - VGEO_Detail inherits from UT_IntrusiveRefCounter now.
      There's no need for virtual decref() and crazy ref count
      locking.
    - Made the segment attributes on VGEO detached attributes.
    - Provide a new interface to create segment attributes without
      having an actual VGEO_Detail, and to append them after the fact.
      This was required by the new procedural interface.
    - Changed VGEO_DetailStash to store VRAY_DetailHandle instead of
      manually managing ref-counts.
    - Updated procedurals to use new interface
    - Added HDK documentation for changes to VRAY procedural API
    - Changed the detail map in VRAY to store VRAY_DetailHandle's
      instead of manually managing ref-counts.