Committed some work in progress work on multiple stream Ogawa loading. Added the ability to reopen an archive with a different number of streams, but something seems to be preventing this from working correctly (caching in the Alembic library, perhaps).
The new archive code will reopen an Ogawa archive with 4 streams, run
threaded code on it, and then reopen it at the end with 1 stream. This is
to prevent running out of file descriptors when many Alembic files are
open at once. It's disabled currently since running threaded on a single
stream is much slower, and running 4 streams is 4x faster, but runs out of
FDs quickly in some cases.