Change GSTY_SubjectPrim/Point to store a GA_ROHandleS instead of looking up every time that it's required.  This allows the GSTY_SubjectPrimGroup/PointGroup to look up the attributes one time instead of N times.
In the 10% case of the bug, this cuts run time from 41s to 33s.

Removed the default arguments for the c-tor, and added a second c-tor
which properly does the per-subject lookup of the attributes.  Added
comments where that constructor is called to note that there might be
performance issues.