Enabling the new viewport alembic code which was previously only available if HOUDINIX_OGL_ALEMBIC_ARCHIVES=2 was set. This enables:
- threaded ogawa loading (2-3x faster)
- alembic shape caching
- instanced alembic rendering in the viewport
- proper alembic refreshing when the alembic file changes
- much faster alembic updates for animated alembics
- Edit SOP updates of packed alembics

The small mesh consolidation (level 3) is not yet enabled due to several
missing features, and the difficulty with caching it. That will provide another
important boost for scenes with lots of unique, small meshes in different
alembics. There is also some work to be done to avoid re-convexing when a
cached shape is reused.