Fixed the HOM Alembic extensions for the Python 3 build.
The main change was to convert the HOM_Alembic.so Houdini plugin into a
standard Python C extension module (renamed to _alembic_hom_extensions.so).
This change was needed in order to register the _alembic_hom_extensions
module with Python's sys.modules for Python 3.

In Python 2, the _alembic_hom_extensions module was registered when
Py_InitModule() was called by the PY_Py_InitModule() wrapper.  However for
Python 3, the PY_Py_InitModule() wrapper calls PyModule_Create() which instead
creates the module but does not register it.  Registration only occurs
internally in Python when the new module is returned by the
`PyInit_<MODULE_NAME>` hook function declared in the .so file.

Therefore, I replaced the HOMextendLibrary() hook with
PyInit__alembic_hom_extensions() and converted the HOM_Alembic.so
plugin to the _alembic_hom_extensions.so Python extension module.

This fixes the sop/basic/abc/inherit regression test and likely other
Alembic tests for Python 3.