Change CH_Channel::getName() to return a UT_StringHolder. This then required, to preserve performance, for OP_Parmareters::getName() to do so.  And thus the getItemName as well....
Troublingly, a lot were constructing with myName(name, 1); but this is
after the sweep to forbid that constructor, so these should be
properly converted.

This required quite a few helper functions that had read-only strings
take a const char * rather than const UT_String &, as they should
never have taking that in the first place if they don't work with
ownership.